+++ packages/config/package.json
