+++ packages/config/src/chains/astar.ts
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import BN from 'bn.js';
-import { Chain } from '@oak-network/xcm-types';
+import { Chain } from '@oak-network/sdk-types';
 import { assets } from '../assets';
-import { ChainAsset, Weight } from '@oak-network/xcm-types';
+import { ChainAsset, Weight } from '@oak-network/sdk-types';
 
 // Shibuya
 const sbyAsset = new ChainAsset({ asset: assets.sby, isNative: true });
