+++ packages/config/src/chains/mangata.ts
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import BN from 'bn.js';
-import { Chain } from '@oak-network/xcm-types';
+import { Chain } from '@oak-network/sdk-types';
 import { assets } from '../assets';
-import { ChainAsset, Weight } from '@oak-network/xcm-types';
+import { ChainAsset, Weight } from '@oak-network/sdk-types';
 
 // mangata-local
 const mgrAsset = new ChainAsset({ asset: assets.mgr, isNative: true });
