+++ packages/config/src/chains/oak.ts
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import BN from 'bn.js';
-import { Chain } from '@oak-network/xcm-types';
+import { Chain } from '@oak-network/sdk-types';
 import { assets } from '../assets';
-import { ChainAsset, Weight } from '@oak-network/xcm-types';
+import { ChainAsset, Weight } from '@oak-network/sdk-types';
 
 // turing-local
 const turAsset = new ChainAsset({ asset: assets.tur, isNative: true });
