+++ packages/provider/tsconfig.json
@@ -2,9 +2,11 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "rootDir": "src",
-    "outDir": "build"
+    "outDir": "build",
+    "skipLibCheck": true
   },
   "references": [
-    { "path": "../xcm-types" }
+    { "path": "../types" },
+    { "path": "../sdk-types" }
   ]
 }
