+++ tsconfig.json
@@ -15,14 +15,14 @@
       "@oak-network/api-augment/*": ["./packages/api-augment/src/*"],
       "@oak-network/types": ["./packages/types/src"],
       "@oak-network/types/*": ["./packages/types/src/*"],
-      "@oak-network/xcm-types": ["./packages/xcm-types/src"],
-      "@oak-network/xcm-types/*": ["./packages/xcm-types/src/*"],
-      "@oak-network/xcm-config": ["./packages/xcm-config/src"],
-      "@oak-network/xcm-config/*": ["./packages/xcm-config/src/*"],
-      "@oak-network/xcm-provider": ["./packages/xcm-provider/src"],
-      "@oak-network/xcm-provider/*": ["./packages/xcm-provider/src/*"],
-      "@oak-network/xcm-sdk": ["./packages/xcm-sdk/src"],
-      "@oak-network/xcm-sdk/*": ["./packages/xcm-sdk/src/*"]
+      "@oak-network/sdk-types": ["./packages/sdk-types/src"],
+      "@oak-network/sdk-types/*": ["./packages/sdk-types/src/*"],
+      "@oak-network/config": ["./packages/config/src"],
+      "@oak-network/config/*": ["./packages/config/src/*"],
+      "@oak-network/provider": ["./packages/provider/src"],
+      "@oak-network/provider/*": ["./packages/provider/src/*"],
+      "@oak-network/sdk": ["./packages/sdk/src"],
+      "@oak-network/sdk/*": ["./packages/sdk/src/*"]
     }
   }
 }
